Add config_file pillar to configure logstash
The Logstash Debian package has a directory it scans for config files,
which is /etc/logstash/conf.d/*.conf. However, the formula for logstash
provided only a way to configure logstash forwarder. This PR makes it
possible to add your own custom configuration file.

Note: I've discussed this with Frank, and he told me it would probably
be a better idea to serialise the configuration file from a pillar
structure to ruby, so that it's easier to configure. Although I fully
understand and support this idea, I don't think it's feasible right now.
I will, however, attempt to support this later.